cycling_performance/app/config.py
import os
from dotenv import load_dotenv

# Charger les variables d'environnement depuis le fichier .env
load_dotenv()

# Variables d'environnement
SECRET_KEY=os.getenv("SECRET_KEY")
ACCESS_TOKEN_EXPIRE_MINUTES=os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES")
DATABASE_URL=os.getenv("DATABASE_URL")
ALGORITHM=os.getenv("ALGORITHM")
